refactor(orders): merge cancel and refund forms into one reversal form
The two form modules (298 + 305 lines) were twins: same line picker,
same per-line reason/note rows, same validation, same submit mapping.
Once the cancel/refund renames are normalized they differed only in the
reason list and its default, the copy, the mutation payload shape, and
the refund-only amount preview added in PR #62.

order-line-reversal-form.tsx now holds one deep form; CancelOrderForm
and RefundOrderForm remain as shallow config wrappers with unchanged
props, so the order-identity-strip call sites keep working as before.
The refund config passes capsByLineKey, which switches on the per-line
amount preview, the running total, and the amount-bearing submit label.
ADR-0008 names cancel the unpaid per-line twin of refund; the code
shape now matches the decision.
