+++ apps/web/src/features/orders/components/order-identity-strip.tsx
@@ -17,11 +17,13 @@ import {
 } from "@/components/ui/dropdown-menu";
 import { OpenComplaintForm } from "@/features/complaints/components/open-complaint-form";
 import { useOpenComplaintMutation } from "@/features/complaints/hooks/useComplaintMutations";
-import { CancelOrderForm } from "@/features/orders/components/cancel-order-form";
 import { OrderCourierForm } from "@/features/orders/components/order-courier-form";
+import {
+	CancelOrderForm,
+	RefundOrderForm,
+} from "@/features/orders/components/order-line-reversal-form";
 import { OrderPickupEventDialog } from "@/features/orders/components/order-pickup-event-dialog";
 import { PaymentStatusBadge } from "@/features/orders/components/payment-status-badge";
-import { RefundOrderForm } from "@/features/orders/components/refund-order-form";
 import {
 	useCancelOrderMutation,
 	useRefundOrderMutation,
