--- apps/web/src/features/orders/components/refund-order-form.tsx
@@ -1,375 +1,0 @@
-import { allocateRefund, lineKey } from "@fresclean/api/schema";
-import { zodResolver } from "@hookform/resolvers/zod";
-import type { UseMutationResult } from "@tanstack/react-query";
-import { Fragment, useMemo } from "react";
-import {
-	Controller,
-	FormProvider,
-	useFieldArray,
-	useForm,
-	useFormContext,
-	useWatch,
-} from "react-hook-form";
-import { z } from "zod";
-import { SelectField } from "@/components/form/select-field";
-import { Button } from "@/components/ui/button";
-import { Checkbox } from "@/components/ui/checkbox";
-import { Field, FieldError, FieldLabel } from "@/components/ui/field";
-import { Textarea } from "@/components/ui/textarea";
-import type { CreateOrderRefundPayload, OrderRefundReason } from "@/lib/api";
-import { formatRefundReason, REFUND_REASONS } from "@/lib/status";
-import { formatIDRCurrency } from "@/shared/utils";
-
-const refundFormSchema = z
-	.object({
-		items: z.array(
-			z.object({
-				id: z.number(),
-				kind: z.enum(["service", "product"]),
-				selected: z.boolean(),
-				reason: z.enum(REFUND_REASONS),
-				note: z.string().optional(),
-			}),
-		),
-	})
-	.superRefine((data, ctx) => {
-		const selectedCount = data.items.filter((item) => item.selected).length;
-		if (selectedCount === 0) {
-			ctx.addIssue({
-				code: "custom",
-				path: ["items"],
-				message: "Select at least one item to refund.",
-			});
-		}
-		data.items.forEach((item, index) => {
-			if (
-				item.selected &&
-				item.reason === "other" &&
-				!(item.note ?? "").trim()
-			) {
-				ctx.addIssue({
-					code: "custom",
-					path: ["items", index, "note"],
-					message: "Note is required when reason is Other.",
-				});
-			}
-		});
-	});
-
-type RefundFormValues = z.infer<typeof refundFormSchema>;
-
-type RefundOrderMutation = UseMutationResult<
-	unknown,
-	Error,
-	{ orderId: number; payload: CreateOrderRefundPayload },
-	unknown
->;
-
-export interface RefundableServiceOption {
-	id: number;
-	item_code: string | null;
-}
-
-export interface RefundableProductOption {
-	id: number;
-	name: string;
-	qty: number;
-}
-
-interface RefundOrderFormProps {
-	capsByLineKey: Map<string, number>;
-	closeDialog: () => void;
-	orderId: number;
-	refundableProducts: RefundableProductOption[];
-	refundableServices: RefundableServiceOption[];
-	refundMutation: RefundOrderMutation;
-}
-
-export const RefundOrderForm = ({
-	capsByLineKey,
-	closeDialog,
-	orderId,
-	refundableProducts,
-	refundableServices,
-	refundMutation,
-}: RefundOrderFormProps) => {
-	const refundLines = [
-		...refundableServices.map((service) => ({
-			kind: "service" as const,
-			id: service.id,
-			label: service.item_code ?? `Service #${service.id}`,
-		})),
-		...refundableProducts.map((product) => ({
-			kind: "product" as const,
-			id: product.id,
-			label: `${product.name} × ${product.qty}`,
-		})),
-	];
-	const hasBothKinds =
-		refundableServices.length > 0 && refundableProducts.length > 0;
-	const form = useForm<RefundFormValues>({
-		resolver: zodResolver(refundFormSchema),
-		defaultValues: {
-			items: refundLines.map((line) => ({
-				id: line.id,
-				kind: line.kind,
-				selected: false,
-				reason: "damaged",
-				note: "",
-			})),
-		},
-	});
-	const { fields } = useFieldArray({ control: form.control, name: "items" });
-	const watchedItems = useWatch({ control: form.control, name: "items" });
-	// Amounts come from the same allocateRefund the server runs at submit.
-	// Selected lines are allocated together (leftover rupiah shift with the
-	// set); unselected lines preview what they'd refund alone.
-	const previewAmounts = useMemo(() => {
-		const amounts = new Map<string, number>();
-		const items = watchedItems ?? [];
-		const selected = items.filter(
-			(item) =>
-				item.selected &&
-				(capsByLineKey.get(lineKey(item.kind, item.id)) ?? 0) > 0,
-		);
-		if (selected.length > 0) {
-			for (const line of allocateRefund({ capsByLineKey, lines: selected })) {
-				amounts.set(lineKey(line.kind, line.id), line.amount);
-			}
-		}
-		for (const item of items) {
-			const key = lineKey(item.kind, item.id);
-			if (!amounts.has(key)) {
-				const cap = capsByLineKey.get(key) ?? 0;
-				amounts.set(
-					key,
-					cap > 0
-						? allocateRefund({
-								capsByLineKey,
-								lines: [{ id: item.id, kind: item.kind }],
-							})[0].amount
-						: 0,
-				);
-			}
-		}
-		return amounts;
-	}, [watchedItems, capsByLineKey]);
-	const totalRefund = (watchedItems ?? []).reduce(
-		(sum, item) =>
-			item.selected
-				? sum + (previewAmounts.get(lineKey(item.kind, item.id)) ?? 0)
-				: sum,
-		0,
-	);
-	const isPending = refundMutation.isPending;
-	const itemsError = form.formState.errors.items;
-	const itemsRootMessage =
-		itemsError && !Array.isArray(itemsError)
-			? (itemsError as { message?: string }).message
-			: undefined;
-
-	const handleSelectAll = () => {
-		fields.forEach((_, index) => {
-			form.setValue(`items.${index}.selected`, true, { shouldDirty: true });
-		});
-		form.clearErrors("items");
-	};
-
-	const handleClear = () => {
-		fields.forEach((_, index) => {
-			form.setValue(`items.${index}.selected`, false, { shouldDirty: true });
-		});
-	};
-
-	const onSubmit = async (values: RefundFormValues) => {
-		const items = values.items
-			.filter((item) => item.selected)
-			.map((item) => ({
-				...(item.kind === "service"
-					? { order_service_id: item.id }
-					: { order_product_id: item.id }),
-				reason: item.reason,
-				note: item.note?.trim() || undefined,
-			}));
-
-		await refundMutation.mutateAsync({
-			orderId,
-			payload: {
-				items,
-			},
-		});
-		closeDialog();
-	};
-
-	return (
-		<FormProvider {...form}>
-			<form
-				className="flex flex-col gap-4"
-				onSubmit={form.handleSubmit(onSubmit)}
-			>
-				<div className="flex flex-wrap items-center justify-between gap-2">
-					<p className="text-muted-foreground text-sm">
-						Select items to refund and choose a reason for each.
-					</p>
-					<div className="flex gap-2">
-						<Button
-							type="button"
-							variant="outline"
-							size="sm"
-							onClick={handleSelectAll}
-							disabled={isPending || fields.length === 0}
-						>
-							Select all
-						</Button>
-						<Button
-							type="button"
-							variant="outline"
-							size="sm"
-							onClick={handleClear}
-							disabled={isPending || fields.length === 0}
-						>
-							Clear
-						</Button>
-					</div>
-				</div>
-
-				{itemsRootMessage ? (
-					<p className="text-destructive text-xs">{itemsRootMessage}</p>
-				) : null}
-
-				<div className="grid max-h-[50vh] gap-3 overflow-y-auto pr-1">
-					{fields.map((field, index) => {
-						const line = refundLines[index];
-						const showKindHeader =
-							hasBothKinds &&
-							(index === 0 || refundLines[index - 1]?.kind !== line?.kind);
-
-						return (
-							<Fragment key={field.id}>
-								{showKindHeader ? (
-									<p className="text-muted-foreground text-xs font-medium uppercase tracking-wide">
-										{line?.kind === "product" ? "Products" : "Services"}
-									</p>
-								) : null}
-								<RefundItemRow
-									amount={
-										line
-											? (previewAmounts.get(lineKey(line.kind, line.id)) ?? 0)
-											: 0
-									}
-									index={index}
-									label={line?.label ?? `Item #${index + 1}`}
-									disabled={isPending}
-								/>
-							</Fragment>
-						);
-					})}
-				</div>
-
-				<div className="flex items-center justify-between border-t pt-3 text-sm">
-					<span className="text-muted-foreground">Refund total</span>
-					<span className="font-medium font-mono tabular-nums">
-						{formatIDRCurrency(String(totalRefund)) || "Rp0"}
-					</span>
-				</div>
-
-				<div className="flex flex-col-reverse gap-2 sm:flex-row sm:justify-end">
-					<Button
-						type="button"
-						variant="outline"
-						onClick={closeDialog}
-						disabled={isPending}
-					>
-						Go back
-					</Button>
-					<Button type="submit" variant="destructive" disabled={isPending}>
-						{isPending
-							? "Refunding…"
-							: totalRefund > 0
-								? `Refund ${formatIDRCurrency(String(totalRefund))}`
-								: "Confirm refund"}
-					</Button>
-				</div>
-			</form>
-		</FormProvider>
-	);
-};
-
-interface RefundItemRowProps {
-	amount: number;
-	disabled: boolean;
-	index: number;
-	label: string;
-}
-
-const RefundItemRow = ({
-	amount,
-	disabled,
-	index,
-	label,
-}: RefundItemRowProps) => {
-	const { control, register } = useFormContext<RefundFormValues>();
-	const selected =
-		useWatch({ control, name: `items.${index}.selected` }) ?? false;
-	const inputsDisabled = disabled || !selected;
-	const checkboxId = `refund-item-${index}`;
-
-	return (
-		<div className="grid gap-2 border p-3">
-			<Controller
-				control={control}
-				name={`items.${index}.selected`}
-				render={({ field }) => (
-					<Field orientation="horizontal">
-						<Checkbox
-							id={checkboxId}
-							checked={field.value}
-							onCheckedChange={(value) => field.onChange(Boolean(value))}
-							disabled={disabled}
-						/>
-						<FieldLabel htmlFor={checkboxId}>{label}</FieldLabel>
-						<span className="ml-auto font-mono text-sm tabular-nums">
-							{formatIDRCurrency(String(amount)) || "Rp0"}
-						</span>
-					</Field>
-				)}
-			/>
-
-			<Controller
-				control={control}
-				name={`items.${index}.reason`}
-				render={({ field }) => (
-					<SelectField
-						items={REFUND_REASONS.map((reason) => ({
-							value: reason,
-							label: formatRefundReason(reason),
-						}))}
-						value={field.value}
-						onValueChange={(value) =>
-							field.onChange(value as OrderRefundReason)
-						}
-						disabled={inputsDisabled}
-						placeholder="Select reason"
-						className="w-full"
-					/>
-				)}
-			/>
-
-			<Controller
-				control={control}
-				name={`items.${index}.note`}
-				render={({ fieldState }) => (
-					<Field data-invalid={fieldState.invalid}>
-						<Textarea
-							placeholder="Reason note"
-							disabled={inputsDisabled}
-							aria-invalid={fieldState.invalid}
-							{...register(`items.${index}.note`)}
-						/>
-						<FieldError errors={[fieldState.error]} />
-					</Field>
-				)}
-			/>
-		</div>
-	);
-};
